Refactor config and validation, fix #12
This commit refactors the entire configuration part of the code: it is
no longer stored in a central place, instead, it is its own package. The
struct also now contains a Validate() function that can run for any
configuration object, making it more flexible and cleaner.

It also changes how validation is done, by improving the performance of
the clients parsing and validation.

Further, it improves how some validation is done, and now net.IP objects
are used for IP Address handling, as opposed to strings before.

Finally, this commit makes the first small change towards a new backend
architecture that is more flexible and allows for adding more backends,
which will be coming in the near future. There is still some refactoring
work to be done in order to be able to fully support them, but the start
has been made.