Large refactor of storage backends
With this commit we are getting one step closer to the support of
multiple backends in eldim. This commit adds backend interfaces to
clearly define what is expected of a backend service from eldim.

It also converts the existing OpenStack Swift backend to use this new
interface binding.

Finally, it modifies all the code to be OpenStack Swift agnostic, with
the exception of course of the relevant backend code.

During this conversion, some Prometheus metrics have been lost, and some
logging of backend-specific things have been removed.