Add support for S3
This commit adds support for the S3 protocol to eldim, as a backend, so
it can now connect to almost every object storage system available.

This fixes #8, and since Backblaze added an S3 compatibility layer on
their B2 product, it also fixes #7.

Three examples for three types of common S3 service providers have been
added to `eldim.yml' to serve as examples.